change mstime_t from  to

+++ src/utils/millisecond_time.h
@@ -1,7 +1,9 @@
 #ifndef MILLISECONDS_TIME_H
 #define MILLISECONDS_TIME_H
 
-typedef long long mstime_t;
+#include <inttypes.h>
+
+typedef uint64_t mstime_t;
 
 /*
  * @return current time in milliseconds
